[S06P12D203-14] [AOS-김주환] feat: 다이얼로그 연동 및 id로 지어 가져오는 API 삭제

+++ aos/Near/app/src/main/java/com/ssafy/near/dto/FingerSignInfo.kt
@@ -3,7 +3,7 @@ package com.ssafy.near.dto
 data class FingerSignInfo(
     val category: Int,
     val explanation: String,
-    val fingercotent_key: Int,
+    val fingercotent_key: Long,
     val image_path: String,
     val name: String,
 )
